lexer.rsを7つのサブモジュールに分割 #75
lexer.rs (2132行) を機能単位のサブモジュールに分割する。

- lexer/note.rs     音符・休符・音長・オクターブ・ゲート・ベロシティ・タイミング・ループ
- lexer/command.rs  read_upper_command のディスパッチと各種コマンド
- lexer/calc.rs     計算式、If/For/While
- lexer/variable.rs 変数・ユーザー定義関数
- lexer/cc.rs       CC / RPN / NRPN / PitchBend / SysEx / 音色
- lexer/args.rs     引数の読み取り
- lexer/error.rs    エラー・警告ログ

lexer.rs は lex() の文字ディスパッチのみを残して227行になった。

実装上の判断:
- src/lexer.rs を親モジュールとして残し src/lexer/*.rs を追加する形にした。
  mod.rs へのリネームを行わないため git blame が保たれる。
- 移動した関数の可視性は pub(super) とし、lexer 外部からは参照できない。
- 親が use xxx::*; で再エクスポートし各子が use super::*; で受けるため、
  呼び出し側のコードは変更していない。

このコミットは移動のみで振る舞いは変更していない。
cargo test 102件通過、ゴールデンMIDIのハッシュも一致。
分割前後の全行を正規化して照合し完全一致を確認済み。
